backend/corporate-retreat-agent/requirements.txt
openai==1.54.3
tavily-python==0.5.0
stripe==11.1.1
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
fastapi==0.115.4
uvicorn==0.32.0
httpx==0.27.2
rich==13.9.4
firecrawl-py==4.14.0
beautifulsoup4==4.12.3
requests==2.31.0
crewai==0.86.0
crewai-tools>=0.17.0